Features

+++ backend/src/main/java/com/example/demo/model/Note.java
@@ -1,7 +1,6 @@
 package com.example.demo.model;
 
 import jakarta.persistence.*;
-import java.math.BigDecimal;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 
 @Entity
@@ -27,7 +26,7 @@ public class Note {
     @JsonBackReference
     private Matiere matiere;
 
-    private BigDecimal valeur;
+    private Double valeur;
 
     public Note(){}
 
@@ -45,7 +44,7 @@ public class Note {
 
     public void setMatiere(Matiere matiere){ this.matiere = matiere; }
 
-    public BigDecimal getValeur(){ return valeur; }
+    public Double getValeur(){ return valeur; }
 
-    public void setValeur(BigDecimal valeur){ this.valeur = valeur; }
+    public void setValeur(Double valeur){ this.valeur = valeur; }
 }
