+++ backend/src/main/java/com/example/demo/model/NoteFinale.java
@@ -1,7 +1,6 @@
 package com.example.demo.model;
 
 import jakarta.persistence.*;
-import java.math.BigDecimal;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 
 @Entity
@@ -13,30 +12,45 @@ public class NoteFinale {
     private Long id;
 
     @ManyToOne
-    @JoinColumn(name="etudiant_id")
+    @JoinColumn(name = "etudiant_id")
     @JsonBackReference
     private Etudiant etudiant;
 
     @ManyToOne
-    @JoinColumn(name="matiere_id")
+    @JoinColumn(name = "matiere_id")
     @JsonBackReference
     private Matiere matiere;
 
-    private BigDecimal valeur;
+    private Double valeur;
 
-    public NoteFinale(){}
+    public NoteFinale() {
+    }
 
-    public Long getId(){ return id; }
+    public Long getId() {
+        return id;
+    }
 
-    public Etudiant getEtudiant(){ return etudiant; }
+    public Etudiant getEtudiant() {
+        return etudiant;
+    }
 
-    public void setEtudiant(Etudiant etudiant){ this.etudiant = etudiant; }
+    public void setEtudiant(Etudiant etudiant) {
+        this.etudiant = etudiant;
+    }
 
-    public Matiere getMatiere(){ return matiere; }
+    public Matiere getMatiere() {
+        return matiere;
+    }
 
-    public void setMatiere(Matiere matiere){ this.matiere = matiere; }
+    public void setMatiere(Matiere matiere) {
+        this.matiere = matiere;
+    }
 
-    public BigDecimal getValeur(){ return valeur; }
+    public Double getValeur() {
+        return valeur;
+    }
 
-    public void setValeur(BigDecimal valeur){ this.valeur = valeur; }
+    public void setValeur(Double valeur) {
+        this.valeur = valeur;
+    }
 }
